Fixed recursive bot startup - faster load times!
So, this is a bit of an interesting one. When you imported main for whatever reason you had to import main, it never checked if it was the original version of main. So it would reload the entirety of the plugins only for that plugin and reconnect to Discord for no good reason. This is a fix for that.